feat(client): create seperate components for entities diagram
